+++ src/main/java/ru/yandex/practicum/filmorate/model/User.java
@@ -24,5 +24,5 @@ public class User {
     @Past(message = "Дата рождения не может быть в будущем")
     private LocalDate birthday;
 
-    private final Set<Integer> friends = new HashSet<>();
+    private final Set<Long> friends = new HashSet<>();
 }
